requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.9
django-session-timeout==0.1.0
filelock==3.18.0
idna==3.10
pillow==11.2.1
platformdirs==4.3.7
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
